+++ TelegramBotAPI-extensions-utils/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/utils/updates/CommandsFilters.kt
@@ -9,13 +9,13 @@ import kotlinx.coroutines.flow.filter
 
 fun <T : BaseSentMessageUpdate> Flow<T>.filterExactCommands(
     commandRegex: Regex
-) = asContentMessages().onlyTextContentMessages().filter { contentMessage ->
+) = asContentMessagesFlow().onlyTextContentMessages().filter { contentMessage ->
     (contentMessage.content.fullEntitiesList().singleOrNull() as? BotCommandTextSource) ?.let { commandRegex.matches(it.command) } == true
 }
 
 fun <T : BaseSentMessageUpdate> Flow<T>.filterCommandsInsideTextMessages(
     commandRegex: Regex
-) = asContentMessages().onlyTextContentMessages().filter { contentMessage ->
+) = asContentMessagesFlow().onlyTextContentMessages().filter { contentMessage ->
     contentMessage.content.fullEntitiesList().any {
         (it as? BotCommandTextSource) ?.let { commandRegex.matches(it.command) } == true
     }